fix(styleobserver): if statement for numeric check was wrong
The previous if statement check would break things if the value was undefined, as such, indexOf which was being checked for would not exist on an undefined/null value. The better check to do is just ensure that value exists as this seems to be a better way of handling this numeric edge case.