docs(how-it-works): Add sentences about observed props added during binding
Aurelia Binding adds the observing property to source object when it does
not exist for a few of the binding commands. This can be confusing for
state management applications that might have a very large state and do not
define every single property on their initial state. After binding runs
in the view model, their state now has defined properties that were
not on an initial state. I think it is important to document this since
state management (like aurelia-store) are becoming popular.

closes #713